* fix, ISIS/IGP daemon: couple of code edits to increase portability of the   daemon which was found not compiling on OpenBSD. Thanks to Robert Smith   for his support.

+++ src/isis/thread.c
@@ -216,7 +216,8 @@ recent_relative_time (void)
 static unsigned int
 cpu_record_hash_key (struct cpu_thread_history *a)
 {
-  return (uintptr_t) a->func;
+  /* XXX: uintptr_t not portable: changed to "unsigned int *" */
+  return (unsigned int) a->func;
 }
 
 static int
